pyproject.toml
[tool.poetry]
name = "threads"
version = "0.1.0"
description = ""
authors = ["Kilari Teja <kilariteja9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
tetraframework = "^0.0.1"
sourcetypes = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
